requirements.txt
certifi==2019.6.16
chardet==3.0.4
dropbox==9.4.0
future==0.17.1
SQLAlchemy==1.3.8
idna==2.8
iso8601==0.1.12
numpy==1.16.4
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyserial==3.4
python-barcode==0.10.0
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
